fix add/delete/delete-all button (#2074)
* fix add/delete/delete-all button

* fix link authorizations

* fix content type id

* rename

* fix name group context

* remove unused use

+++ BackofficeBundle/Resources/public/ecmascript/OpenOrchestra/Application/View/Site/SiteFormView.js
@@ -2,6 +2,7 @@ import AbstractFormView     from '../../../Service/Form/View/AbstractFormView'
 import Application          from '../../Application'
 import Site                 from '../../Model/Site/Site'
 import FormViewButtonsMixin from '../../../Service/Form/Mixin/FormViewButtonsMixin'
+import ApplicationError     from '../../../Service/Error/ApplicationError'
 
 /**
  * @class SiteFormView
@@ -12,8 +13,9 @@ class SiteFormView extends mix(AbstractFormView).with(FormViewButtonsMixin)
      * Initialize
      * @param {Form}   form
      * @param {Array}  name
+     * @param {string} siteId
      */
-    initialize({form, name, siteId}) {
+    initialize({form, name, siteId = null}) {
         super.initialize({form : form});
         this._name = name;
         this._siteId = siteId;
@@ -34,7 +36,7 @@ class SiteFormView extends mix(AbstractFormView).with(FormViewButtonsMixin)
     }
 
     /**
-     * Redirect to edit user view
+     * Redirect to edit site view
      *
      * @param {mixed}  data
      * @param {string} textStatus
@@ -51,17 +53,17 @@ class SiteFormView extends mix(AbstractFormView).with(FormViewButtonsMixin)
             siteId: siteId,
             name: name
         });
-        let message = new FlashMessage(data, 'success');
-        FlashMessageBag.addMessageFlash(message);
         Backbone.Events.trigger('form:deactivate', this);
         Backbone.history.navigate(url, true);
     }
 
     /**
      * Delete
-     * @param {event} event
      */
-    _deleteElement(event) {
+    _deleteElement() {
+        if (null === this._siteId) {
+            throw new ApplicationError('Invalid siteId');
+        }
         let site = new Site({'site_id': this._siteId});
         site.destroy({
             success: () => {
